suffix all participating organisations
This makes sure all reporting organisations are saved with their IATI
ref as original ref. Previously this was dependent on order of parsing.
That did not matter for org names or organisation refs shown on
front-ends, these were all shown correctly, it could occur on getting
activities by reporting org.

The fix suffixes the org ref of all organizations that are added
through the participating organizations IATI element.
This makes sure all non-suffixed refs are available for reporting
organisations.

The current solution is a quick patch. The current flow should be
evaluated on updating the IATI parser.